+++ onchain/rollups-cli/src/connect.ts
@@ -24,6 +24,8 @@ import {
 } from "@cartesi/rollups";
 import goerli from "@cartesi/rollups/export/abi/goerli.json";
 import polygon_mumbai from "@cartesi/rollups/export/abi/polygon_mumbai.json";
+import arbitrum_goerli from "@cartesi/rollups/export/abi/arbitrum_goerli.json";
+import optimism_goerli from "@cartesi/rollups/export/abi/optimism_goerli.json";
 
 type DeploymentContract = {
     address: string;
@@ -39,6 +41,8 @@ type Deployment = {
 const deployments: Record<number, Deployment> = {
     5: goerli,
     80001: polygon_mumbai,
+    421613: arbitrum_goerli,
+    420: optimism_goerli,
 };
 
 interface RollupsContracts {